# openapi.yaml
openapi: 3.0.0
info:
  title: Card Benefits Admin API
  description: |
    Admin management API for card and benefit administration.
    
    ## Rate Limiting
    All endpoints are rate limited per admin user:
    - List operations: 100-150 requests per minute
    - Create operations: 50 requests per minute
    - Update operations: 50 requests per minute
    - Delete operations: 30 requests per minute
    
    ## Pagination
    All list endpoints support pagination:
    - Default limit: 20-50 items per page (varies by endpoint)
    - Maximum limit: 100 items per page
    - Responses include pagination metadata (total, page, limit, totalPages, hasMore)
    
    ## Authentication
    All endpoints require admin role authentication via cookie (auth_token)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Development
...
paths:
...
  /api/admin/cards/{id}:
    get:
      tags:
        - Cards
      summary: Get card details
      description: Retrieve single card by ID with benefit count
      operationId: getCard
      x-rateLimit:
        limit: 200
        window: minute
        description: 200 requests per minute per admin user
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
      responses:
        "200":
          description: Card details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Card"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []

    patch:
      tags:
        - Cards
      summary: Update card
      description: Update card properties (PATCH - partial update)
      operationId: updateCard
      x-rateLimit:
        limit: 50
        window: minute
        description: 50 requests per minute per admin user
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cardName:
                  type: string
                  maxLength: 200
                defaultAnnualFee:
                  type: integer
                  minimum: 0
                cardImageUrl:
                  type: string
                  format: uri
                isActive:
                  type: boolean
      responses:
        "200":
          description: Card updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Card"
                  message:
                    type: string
                  changes:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        old:
                          type: any
                        new:
                          type: any
        "409":
          description: Duplicate card name/issuer combination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []

    delete:
      tags:
        - Cards
      summary: Delete card
      description: Delete card or archive if in use
      operationId: deleteCard
      x-rateLimit:
        limit: 30
        window: minute
        description: 30 requests per minute per admin user
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
        - name: force
          in: query
          description: Force delete even if in use
          schema:
            type: boolean
            default: false
        - name: archiveInstead
          in: query
          description: Archive instead of delete if in use
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Card deleted/archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      cardName:
                        type: string
                      action:
                        type: string
                        enum:
                          - deleted
                          - archived
                  message:
                    type: string
        "409":
          description: Card in use (cannot delete without force)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
...
components:
  schemas:
    Card:
      type: object
      properties:
        id:
          type: string
        issuer:
          type: string
        cardName:
          type: string
        defaultAnnualFee:
          type: integer
        cardImageUrl:
          type: string
          format: uri
        displayOrder:
          type: integer
        isActive:
          type: boolean
        isArchived:
          type: boolean
        benefitCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HTTP cookie authentication
security:
  - cookieAuth: []